Moved all endpoints in to RadioDNS-standardised path
ETSI standardisation of RadioEPG in to an ETSI specification introduced a versioned path for locating IP-based documents. The RadioVIS part of the Slideshow specification also has a similar path for it's endpoint. By using a path we avoid the potential of conflicting with other endpoints on a multi-purpose server. The inclusion of a version number also makes future development of the specification easier to manage without breaking legacy implementations.