Add optional 'time_source' parameter
There are some concerns about how reliable the time attribute will be from differing clients with differing "clock" sources. It will be difficult to tackle this within the timescales to deliver a finalised 1.0 specification, so as a compromise and to aide future investigation, we have added an optional 'time_source' attribute which allows a client to inform the broadcaster of where they sourced the time from.

This will allow a broadcaster to make a better decision about how reliable that time might be and therefore potenitally vary their response to the incoming tag to increase or decreate the range/scope of returned events.